refactor: move shared utilities out of app/ directory
Move files that were in app/ but imported by components/hooks:

- app/markets/_lib/formatters.ts → lib/market-formatters.ts
  Updated 11 import sites (2 in components, 9 in app/markets/)

- app/agents/create/components/ → components/agents/create/components/
- app/agents/create/hooks/ → components/agents/create/hooks/
  Updated 1 import site (components/agents/AgentCreate.tsx)

This eliminates cross-boundary imports where shared components
imported from the app/ directory, which would break webpack
aliases in the mobile app.
