+++ apps/web/src/app/betting/page.tsx
@@ -19,7 +19,7 @@ import {
 import { useRouter } from 'next/navigation';
 import { useMemo, useState } from 'react';
 import { toast } from 'sonner';
-import { formatPrice } from '@/app/markets/_lib/formatters';
+import { formatPrice } from '@/lib/market-formatters';
 import { PageContainer } from '@/components/shared/PageContainer';
 import { Skeleton } from '@/components/shared/Skeleton';
 import { useAuth } from '@/hooks/useAuth';
