+++ apps/web/src/app/markets/_components/perps-terminal/PerpsMarketListPanel.tsx
@@ -5,7 +5,7 @@ import { ArrowDown, ArrowUp, Filter, Search, Star } from 'lucide-react';
 import { useMemo, useState } from 'react';
 import type { PerpMarket } from '@/types/markets';
 import { useWatchlistStore } from '../../_hooks/useWatchlistStore';
-import { formatPrice, formatVolume } from '../../_lib/formatters';
+import { formatPrice, formatVolume } from '@/lib/market-formatters';
 
 type SortKey =
   | 'ticker'
