+++ apps/web/src/app/markets/_components/terminal/MarketsTradingTerminal.tsx
@@ -84,7 +84,7 @@ import type {
   TradeSide,
 } from '@/types/markets';
 import { MARKET_TIME_RANGES } from '@/types/markets';
-import { formatBalance } from '../../_lib/formatters';
+import { formatBalance } from '@/lib/market-formatters';
 import { PerpsOrderEntryPanel } from '../perps-terminal/PerpsOrderEntryPanel';
 import { useMarketsTutorial } from '../tutorial/useMarketsTutorial';
 import {
