+++ apps/web/src/components/profile/PositionDetailModal.tsx
@@ -20,7 +20,7 @@ import {
 } from 'lucide-react';
 import { useCallback, useEffect, useState } from 'react';
 import { toast } from 'sonner';
-import { formatPrice } from '@/app/markets/_lib/formatters';
+import { formatPrice } from '@/lib/market-formatters';
 import { FollowButton } from '@/components/interactions';
 import { useAuth } from '@/hooks/useAuth';
 import { usePerpMarketsStore } from '@/stores/perpMarketsStore';
